requirements.txt
tensorflow
numpy
pytorch
plotly
seaborn
scipy
scapy
nltk
scikit-learn
pandas
matplotlib
